% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdro_data.R
\name{create_hdro_data}
\alias{create_hdro_data}
\title{Constructor for HDI Data Class}
\usage{
create_hdro_data(data)
}
\arguments{
\item{data}{A data frame containing HDRO indicators.}
}
\value{
An object of class `hdro_data`.
}
\description{
This function creates a new S3 class `hdro_data`.
}
