% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdro_file.R
\name{hdro_data}
\alias{hdro_data}
\title{Constructor for HDRO Data Class}
\usage{
hdro_data(data)
}
\arguments{
\item{data}{A data frame containing HDRO indicators.}
}
\value{
An object of class `hdro_data`.
}
\description{
This function creates a new S3 class `hdro_data`.
}
