% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hdro_data.R
\name{plot.hdro_data}
\alias{plot.hdro_data}
\title{Plot Method for HDRO Data Class}
\usage{
\method{plot}{hdro_data}(x, indicator, ...)
}
\arguments{
\item{x}{An object of class `hdro_data`.}

\item{indicator}{The indicator to be plotted.}

\item{...}{Additional arguments passed to the plot method.}
}
\description{
This function creates a plot for objects of class `hdro_data`.
}
