% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hdro_data.R
\name{print.hdro_data}
\alias{print.hdro_data}
\title{Print Method for HDRO Data Class}
\usage{
\method{print}{hdro_data}(x, ...)
}
\arguments{
\item{x}{An object of class `hdro_data`.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
This function defines how objects of class `hdro_data` are printed
}
\details{
Task 3 of the HDI project
}
