% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdi_data.R, R/read_hdro_file.R
\name{read_hdro_file}
\alias{read_hdro_file}
\title{Read HDRO Indicator File}
\usage{
read_hdro_file(file_path)

read_hdro_file(file_path)
}
\arguments{
\item{file_path}{The path to the HDRO indicators CSV file.}
}
\value{
An object of class `hdro_data` with correctly assigned data types and cleaned column names.

An object of class `hdro_data` with correctly assigned data types and cleaned column names.
}
\description{
This function reads an HDRO indicators CSV file and assigns appropriate data types to the columns.

This function reads an HDRO indicators CSV file and assigns appropriate data types to the columns.
}
\examples{
df <- read_hdro_file("path/to/hdro_indicators_COUNTRYNAME.csv")
df <- read_hdro_file("path/to/hdro_indicators_COUNTRYNAME.csv")
}
