% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandeephdi-package.R
\docType{package}
\name{sandeephdi}
\alias{sandeephdi-package}
\alias{sandeephdi}
\title{sandeephdi: An R Package for HDI Indicator Data
This package was created as part of an assignment.}
\description{
The `sandeephdi` package provides functions to read, summarize, and plot any country's HDI data.
}
\section{Main Function}{

The main function in this package is `read_hdro_file`, which reads an HDRO indicators CSV file and assigns appropriate data types to the columns.
}

\section{Additional Functions}{

The package also provides methods for printing (`print.hdro_data`), summarizing (`summary.hdro_data`), and plotting (`plot.hdro_data`) the data.
}

