% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hdro_data.R
\name{summary.hdro_data}
\alias{summary.hdro_data}
\title{Summary Method for HDRO Data Class}
\usage{
\method{summary}{hdro_data}(object, ...)
}
\arguments{
\item{object}{An object of class `hdro_data`.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
This function provides a summary for objects of class `hdro_data`.
}
\details{
Task 4
}
