% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdro_file.R
\name{hdro_data}
\alias{hdro_data}
\title{Constructor for HDRO Data Class}
\usage{
hdro_data(data)
}
\arguments{
\item{data}{A data frame containing HDI indicators.}
}
\value{
An object of class `hdro_data`.
}
\description{
This function reads an HDI indicators CSV file and assigns appropriate data types to the columns
This is Task 1 of the HDI project
}
